Split sign video editing into it's own file
