+++ app/frontend/components/edit-sign-form.js
@@ -1,55 +1,4 @@
 import { initialHTML, errorHTML } from "./file-upload";
-import uppyFileUpload from "./uppy-file-upload";
-import signVideoRestrictions from "./uppy/signVideoRestrictions";
-import { put } from "@rails/request.js";
-
-const updateSignVideo = (signId, signedBlobId) => {
-  return put(`/signs/${signId}`, {
-    body: JSON.stringify({
-      sign: { video: signedBlobId },
-    }),
-  }).then((response) => {
-    if (!response.ok) {
-      return Promise.reject(
-        new Error("Something went wrong creating this sign.")
-      );
-    }
-  });
-};
-
-const updateSignVideoController = (container) => {
-  const signId = container.dataset.updateSignVideoSignIdValue;
-
-  if (!signId) {
-    console.error(
-      "Missing required value [data-update-sign-video-sign-id-value]"
-    );
-    return;
-  }
-
-  const uppy = uppyFileUpload(container, {
-    uppy: { restrictions: signVideoRestrictions },
-    dashboard: {
-      hideRetryButton: true,
-      inline: false,
-      trigger: container.querySelector(
-        "[data-update-sign-video-target='trigger']"
-      ),
-    },
-  });
-
-  uppy.addPostProcessor(([fileId]) => {
-    const file = uppy.getFile(fileId);
-    uppy.emit("postprocess-progress", file, {
-      mode: "indeterminate",
-      message: "Creating signs...",
-    });
-
-    updateSignVideo(signId, file.response.signed_id).then(() => {
-      window.location.reload();
-    });
-  });
-};
 
 const handleSignAttachmentUpload = ($container, listSelector, path) => {
   const $content = $container.find(".file-upload__content");
@@ -119,10 +68,4 @@ $(document).on("blur keydown", ".js-attachment-description", (event) => {
     method: "PATCH",
     data: { format: "js", description: input.value },
   });
-});
-
-$(() => {
-  $("[data-controller='update-sign-video']").each((_idx, container) =>
-    updateSignVideoController(container)
-  );
 });
